+++ src/MarkdownConverter.WebApp/wwwroot/service-worker.published.js
@@ -6,14 +6,16 @@ self.addEventListener('fetch', event => event.respondWith(onFetch(event)));
 
 const cacheNamePrefix = 'offline-cache-';
 const cacheName = `${cacheNamePrefix}${self.assetsManifest.version}`;
-const offlineAssetsInclude = [/\.dll$/, /\.pdb$/, /\.wasm/, /\.html/, /\.js$/, /\.json$/, /\.css$/, /\.woff$/, /\.png$/, /\.jpe?g$/, /\.gif$/, /\.ico$/, /\.blat$/, /\.dat$/];
+const offlineAssetsInclude = [/\.dll$/, /\.pdb$/, /\.wasm/, /\.html/, /\.js$/, /\.json$/, /\.css$/, /\.woff$/, /\.png$/, /\.jpe?g$/, /\.gif$/, /\.ico$/, /\.blat$/, /\.dat$/, /\.webmanifest$/];
 const offlineAssetsExclude = [/^service-worker\.js$/];
+const scopeUrl = new URL(self.registration.scope);
+const manifestUrlList = self.assetsManifest.assets.map(asset => new URL(asset.url, scopeUrl).href);
 
 async function onInstall(event) {
     const assetsRequests = self.assetsManifest.assets
         .filter(asset => offlineAssetsInclude.some(pattern => pattern.test(asset.url)))
         .filter(asset => !offlineAssetsExclude.some(pattern => pattern.test(asset.url)))
-        .map(asset => new Request(asset.url, { integrity: asset.hash, cache: 'no-cache' }));
+        .map(asset => new Request(new URL(asset.url, scopeUrl), { integrity: asset.hash, cache: 'no-cache' }));
     const cache = await caches.open(cacheName);
     await cache.addAll(assetsRequests);
 }
@@ -26,8 +28,16 @@ async function onActivate(event) {
 }
 
 async function onFetch(event) {
-    if (event.request.method !== 'GET') return fetch(event.request);
-    const cache = await caches.open(cacheName);
-    const cachedResponse = await cache.match(event.request);
+    let cachedResponse = null;
+    if (event.request.method === 'GET') {
+        const shouldServeIndexHtml = event.request.mode === 'navigate'
+            && !manifestUrlList.some(url => url === event.request.url);
+        const request = shouldServeIndexHtml
+            ? new Request(new URL('index.html', scopeUrl))
+            : event.request;
+        const cache = await caches.open(cacheName);
+        cachedResponse = await cache.match(request);
+    }
+
     return cachedResponse || fetch(event.request);
 }